requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0

# Image/Video Processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# AI Models
realesrgan>=0.3.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8

# Model Conversion
onnx>=1.14.0
coremltools>=7.0

# Web Interface
gradio>=3.40.0
streamlit>=1.25.0

# CLI & Utils
rich>=13.0.0
tqdm>=4.65.0
click>=8.1.0

# Benchmarking & Profiling
psutil>=5.9.0
py-cpuinfo>=9.0.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.3.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Optional: For advanced features
# scikit-image>=0.21.0
# imagehash>=4.3.1